Make white-list of proxy IP addresses optional
While it's a good idea to have a white-list of IP addresses of proxy servers and check against it, it may be difficult to use it in a dynamic context such as Docker and Kubernetes. If the proxy server is part of the Docker and/or Kubernetes environment the IP address will not be fix. Additionally the list of possible IP addresses of the proxy servers might be huge. Here a check should against the white-list be optional.